[PLG-1450]: Generalize subscription creation endpoint

+++ 935-subscription-sdk/src/main/java/io/harness/subscription/helpers/StripeHelper.java
@@ -30,7 +30,7 @@ public interface StripeHelper {
 
   PriceCollectionDTO getPrices(ModuleType moduleType);
   Price getPrice(ModuleType moduleType, String type, String edition, String paymentFrequency);
-  Price getPrice(ModuleType moduleType, String type, String edition, String paymentFrequency, int quantity);
+  Price getPrice(ModuleType moduleType, String type, String edition, String paymentFrequency, long quantity);
   Price getPrice(String lookupKey);
   PriceCollectionDTO listPrices(List<String> lookupKeys);
   SubscriptionDetailDTO createSubscription(SubscriptionParams subscriptionParams);
